+++ index.js
@@ -4,14 +4,9 @@ const figlet = require('figlet');
 const { statSync } = require('fs');
 const { getFiles } = require('./readDir');
 const { readFiles, readFilePrueba } = require('./readFile');
-const { requestStatus, request } = require('./httpRequest.js')
+const { requestStatus } = require('./httpRequest.js')
 
 const dir = process.argv[2];
-
-
-const mdLinks = (path, options) => {
-
-}
 
 
 
@@ -24,15 +19,16 @@ if (statSync(dir).isDirectory()) {
   console.log('Requesting links status: -------------------------------------')
     //const objStatus = requestStatus(parsedData)
     //console.log(Promise.all(objStatus));
-  request(parsedData)
-    .then((status) => {
-      console.log(status);
-    })
+  requestStatus(parsedData).then((status) => {
+    console.log(status);
+  })
 
 } else if (statSync(dir).isFile()) {
   console.log('es un archivo')
   const parsedData = readFilePrueba(dir)
-  request(parsedData)
+  requestStatus(parsedData).then((status) => {
+    console.log(status);
+  })
 } else {
   console.log('argumentos inválidos')
 }
