+++ validate.js
@@ -1,5 +1,5 @@
-const validate = (obj) => {
-  return new Promise((resolve, reject) => {
+const { httpRequest } = require('./httpRequest')
 
-  })
+const validate = (obj) => {
+  return new Promise((resolve, reject) => {})
 }
